Enemy Spawner coded
The feature is not finished. At the moment the enemies aren't stored in an array, they are just spawned in their respective positions.

There is currently a bug in the Projectile/Enemy interaction where, if multiple projectiles are on the screen and the queue_free() is called on the projectiles, other projectiles are "Deactivated"